refactor(workspace-host): move resource-action coordination into ResourceActionExecutor
- Move queue/abort-controller Maps and 5 coordination methods from WorkspaceService into ResourceActionExecutor
- Add dispose() method with disposed guard to ResourceActionExecutor
- Keep delegating runResourceAction on WorkspaceService for API stability
- Keep switchWorktreeEnvironment in WorkspaceService (depends on runLifecycleSetup)
- Redirect 4 cleanup call sites through executor
- Update 3 test bracket-notation accesses to chain through public executor
